Change logic of records of current and last month formatted into useGuestUser
